fix: address code comments and cleanup
