+++ src/providers/google-vertex-ai/types.ts
@@ -139,3 +139,60 @@ export interface GoogleSearchRetrievalTool {
     };
   };
 }
+
+type GoogleBatchJobStatus =
+  | 'JOB_STATE_UNSPECIFIED'
+  | 'JOB_STATE_QUEUED'
+  | 'JOB_STATE_PENDING'
+  | 'JOB_STATE_RUNNING'
+  | 'JOB_STATE_SUCCEEDED'
+  | 'JOB_STATE_FAILED'
+  | 'JOB_STATE_CANCELLING'
+  | 'JOB_STATE_CANCELLED'
+  | 'JOB_STATE_PAUSED'
+  | 'JOB_STATE_EXPIRED'
+  | 'JOB_STATE_UPDATING'
+  | 'JOB_STATE_PARTIALLY_SUCCEEDED';
+
+export interface GoogleBatchRecord {
+  /**
+   * @example projects/562188160088/locations/us-east4/batchPredictionJobs/{id}
+   */
+  name: string;
+  displayName: string;
+  /**
+   * @example projects/562188160088/locations/us-east4/models/{model}
+   */
+  model: string;
+  inputConfig: {
+    instancesFormat: 'jsonl';
+    gcsSource: {
+      uris: string;
+    };
+  };
+  outputConfig: {
+    predictionsFormat: 'jsonl';
+    gcsDestination: {
+      outputUriPrefix: string;
+    };
+  };
+  outputInfo?: {
+    gcsOutputDirectory: string;
+  };
+  state: GoogleBatchJobStatus;
+  createTime: string;
+  updateTime: string;
+  modelVersionId: string;
+  error?: {
+    code: string;
+    message: string;
+  };
+  startTime: string;
+  endTime: string;
+  completionsStats?: {
+    successfulCount: string;
+    failedCount: string;
+    incompleteCount: string;
+    successfulForecastPointCount: string;
+  };
+}
