+++ src/providers/types.ts
@@ -358,92 +358,8 @@ export interface FinetuneRequest {
   };
 }
 
-export interface BedrockFinetuneRecord {
-  baseModelArn: string;
-  creationTime: string;
-  customModelArn: string;
-  customModelName: string;
-  customizationType: string;
-  endTime: string;
-  jobArn: string;
-  jobName: string;
-  lastModifiedTime: string;
-  status: 'Completed' | 'Failed' | 'InProgress' | 'Stopping' | 'Stopped';
-  failureMessage?: string;
-  validationDataConfig?: {
-    s3Uri: string;
-  };
-  trainingDataConfig?: {
-    s3Uri: string;
-  };
-  hyperParameters?: {
-    learningRate: number;
-    batchSize: number;
-    epochCount: number;
-  };
-  outputModelName?: string;
-  outputModelArn?: string;
-}
-
 export interface CreateBatchRequest {
   input_file_id: string;
   endpoint: string;
   completion_window: string;
-}
-
-type GoogleBatchJobStatus =
-  | 'JOB_STATE_UNSPECIFIED'
-  | 'JOB_STATE_QUEUED'
-  | 'JOB_STATE_PENDING'
-  | 'JOB_STATE_RUNNING'
-  | 'JOB_STATE_SUCCEEDED'
-  | 'JOB_STATE_FAILED'
-  | 'JOB_STATE_CANCELLING'
-  | 'JOB_STATE_CANCELLED'
-  | 'JOB_STATE_PAUSED'
-  | 'JOB_STATE_EXPIRED'
-  | 'JOB_STATE_UPDATING'
-  | 'JOB_STATE_PARTIALLY_SUCCEEDED';
-
-export interface GoogleBatchRecord {
-  /**
-   * @example projects/562188160088/locations/us-east4/batchPredictionJobs/{id}
-   */
-  name: string;
-  displayName: string;
-  /**
-   * @example projects/562188160088/locations/us-east4/models/{model}
-   */
-  model: string;
-  inputConfig: {
-    instancesFormat: 'jsonl';
-    gcsSource: {
-      uris: string;
-    };
-  };
-  outputConfig: {
-    predictionsFormat: 'jsonl';
-    gcsDestination: {
-      outputUriPrefix: string;
-    };
-  };
-  outputInfo?: {
-    gcsOutputDirectory: string;
-  };
-  state: GoogleBatchJobStatus;
-  createTime: string;
-  updateTime: string;
-  modelVersionId: string;
-  error?: {
-    code: string;
-    message: string;
-  };
-  startTime: string;
-  endTime: string;
-  completionsStats?: {
-    successfulCount: string;
-    failedCount: string;
-    incompleteCount: string;
-    successfulForecastPointCount: string;
-  };
 }
